statics: add demo image
Note that we are not parsing the images to make it smaller in size
or to add a hash to its name, but this configuration should be easy
to do.

For now, the assets you put into `/static` will be just moved to
`/out/static`, but it is preferable to use imagemin and hash
path names for a proper cache invalidation.